organiza o reposiório, define a documentação inicial e adiciona o wokflow
