+++ docs/docs/interpretador/comoRodar.md
@@ -1,0 +1,3 @@
+# Como rodar o projeto
+
+EXPLICAÇÃO DE COMO EXECUTAR O PROJETO
