+++ docs/docs/interpretador/glossarioTermos.md
@@ -1,0 +1,42 @@
+# Glossário de Termos
+**Subconjunto de C a ser suportado**
+
+Este interpretador implementa uma versão reduzida do **C**, focada nas construções essenciais para compreensão e prática de compiladores. Não implementa todas as funcionalidades complexas do C padrão.
+
+
+### Tipos básicos
+- `int`  
+- `float`  
+- `char`  
+- `double`  
+- `void`
+
+
+### Operações Aritméticas
+- Soma (`+`)  
+- Subtração (`-`)  
+- Multiplicação (`*`)  
+- Divisão (`/`)  
+- Módulo (`%`)
+
+
+### Controle de Fluxo
+- `if / else`  
+- `while`  
+- `for`
+
+
+### Variáveis e Atribuições
+- Declarações simples: `int x;`, `float y;`  
+- Atribuições: `x = 5;`  
+
+
+### Estruturas Suportadas
+- Blocos básicos: `{ ... }`  
+- Leitura e escrita simples usando **entrada e saída padrão** (`scanf`, `printf`)
+
+# Histórico de versões
+
+| Versão | Data       | Autor           | Descrição da Alteração                  | Revisor   |
+|--------|-----------|----------------|----------------------------------------|--------------|
+| 1.0    | 13/09/2025 | Ana Catarina    | Adição inicial do glossário de termos | |
