+++ docs/docs/sprints/organizacao.md
@@ -1,0 +1,41 @@
+# Organização e fluxo do trabalho
+
+Esta seção descreve a rotina do time, práticas adotadas, comunicação, versionamento e documentação do projeto, garantindo organização, transparência e produtividade durante o desenvolvimento do interpretador.
+
+## 1. Reuniões Semanais
+
+- **Dia/Horário:** Quartas-feiras, 14h  
+- **Tipo:** Alinhamento semanal  
+- **Pauta:**
+  - Progresso das tarefas e sprints
+  - Bloqueadores enfrentados
+  - Planejamento da próxima sprint
+  - Dúvidas e esclarecimentos com o professor
+
+> Observação: Registrar decisões e pontos importantes no comentário das issues correspondentes.
+
+## 2. Comunicação Assíncrona
+
+- Utilizar **WhatsApp** ou ferramenta de mensagens do time para alinhamentos rápidos.  
+- **Documentar evolução e bloqueadores** diretamente nos comentários das issues para manter registro e histórico do desenvolvimento.  
+- Garantir que todos os membros estejam cientes das atualizações importantes.
+
+## 3. Versionamento de Código
+
+- Criar **uma branch por feature ou atividade**.  
+- Branches devem estar **linkadas às suas respectivas issues**.  
+- Usar **nomes significativos** para as branches (ex.: `sprint1-lex-parser`).  
+- **Pull requests** devem ser revisados por outro membro do time antes do merge na `main`.  
+- Manter um **histórico claro de commits** e alterações significativas.
+
+## 4. Documentação
+
+- Atualização contínua durante o desenvolvimento.  
+- Manter **histórico de versão** das entregas e documentos.  
+- Registrar mudanças importantes no **Histórico de Versão**.
+
+# Histórico de versões
+
+| Versão | Data       | Autor           | Descrição da Alteração                  | Revisor   |
+|--------|-----------|----------------|----------------------------------------|--------------|
+| 1.0    | 13/09/2025 | Ana Catarina    | Adição do fluxo de trabalho | |
