+++ docs/docs/sprints/sprint1/cronograma.md
@@ -1,0 +1,34 @@
+# Sprint 1
+
+| Sprint | |
+|--------|----------------|
+| Início | 10/09/2025 |
+| Término | 18/09/2025 |
+| Duração | 8 dias |
+
+## Objetivo
+Configurar o projeto, definir a linguagem e criar os protótipos iniciais (`lex.l` e `parser.y`) para dar início ao desenvolvimento do interpretador.
+
+## Cronograma planejado
+
+| Issue | Atividade | Descrição | Início | Término | Responsável |
+|-------|-----------|-----------|--------|---------|-------------|
+| - | Criar o repositório | Criar o repositório do projeto, configurar e organizar a documentação | 10/09/2025 | 13/09/2025 | Igor, Ana Catarina |
+| [1](https://github.com/igorvdaniel/Trabalho_Final_Compiladores1_2025-02/issues/1) | Card: Definir e Documentar a Linguagem | Criar `docs/language_spec.md` com tipos suportados, operadores, estruturas de controle e exemplos | 13/09/2025 | 13/09/2025 | Ana Catarina |
+| [2](https://github.com/igorvdaniel/Trabalho_Final_Compiladores1_2025-02/issues/2) | Card: Criar arquivo lex.l (Flex) | Definir tokens básicos: identificadores, números, operadores e palavras-chave | 13/09/2025 | 16/09/2025 | Ana Catarina |
+| [3](https://github.com/igorvdaniel/Trabalho_Final_Compiladores1_2025-02/issues/3) | Card: Criar arquivo parser.y (Bison) | Criar protótipo inicial da gramática com regras básicas e incompletas | 13/09/2025 | 16/09/2025 | Igor |
+
+## Cronograma executado
+
+| Issue | Atividade | Descrição | Início | Término | Responsável | Revisor |
+|-------|-----------|-----------|--------|---------|-------------|---------|
+| - | Criar o repositório | Criar o repositório do projeto, configurar e organizar a documentação | 10/09/2025 | 13/09/2025 | Igor, Ana Catarina | Ana Catarina |
+| [1](https://github.com/igorvdaniel/Trabalho_Final_Compiladores1_2025-02/issues/1) | Card: Definir e Documentar a Linguagem | Criar `docs/language_spec.md` com tipos suportados, operadores, estruturas de controle e exemplos | 13/09/2025 | 13/09/2025 | Ana Catarina | Igor |
+| [2](https://github.com/igorvdaniel/Trabalho_Final_Compiladores1_2025-02/issues/2) | Card: Criar arquivo lex.l (Flex) | Definir tokens básicos: identificadores, números, operadores e palavras-chave | 13/09/2025 | 16/09/2025 | Ana Catarina | Igor |
+| [3](https://github.com/igorvdaniel/Trabalho_Final_Compiladores1_2025-02/issues/3) | Card: Criar arquivo parser.y (Bison) | Criar protótipo inicial da gramática com regras básicas e incompletas | 13/09/2025 | 16/09/2025 | Igor | Ana Catarina |
+
+# Histórico de versões
+
+| Versão | Data       | Autor           | Descrição da Alteração                  | Revisor   |
+|--------|-----------|----------------|----------------------------------------|--------------|
+| 1.0    | 13/09/2025 | Ana Catarina    | Adição do cronograma da sprint 1 | |
