+++ docs/docs/sprints/sprint2/cronograma.md
@@ -1,0 +1,7 @@
+# Sprint 2
+
+| Sprint | |
+|--------|----------------|
+| Início | 19/09/2025 |
+| Término | 02/10/2025 |
+| Duração | 13 dias |
