+++ docs/docs/sprints/sprint3/cronograma.md
@@ -1,0 +1,7 @@
+# Sprint 3
+
+| Sprint | |
+|--------|----------------|
+| Início | 03/10/2025 |
+| Término | 16/10/2025 |
+| Duração | 13 dias |
