+++ docs/docs/sprints/sprint4/cronograma.md
@@ -1,0 +1,7 @@
+# Sprint 4
+
+| Sprint | |
+|--------|----------------|
+| Início | 17/10/2025 |
+| Término | 30/10/2025 |
+| Duração | 13 dias |
