+++ docs/docs/sprints/sprint5/cronograma.md
@@ -1,0 +1,7 @@
+# Sprint 5
+
+| Sprint | |
+|--------|----------------|
+| Início | 31/10/2025 |
+| Término | 20/11/2025 |
+| Duração | 21 dias |
