+++ docs/docs/sprints/sprint6/cronograma.md
@@ -1,0 +1,7 @@
+# Sprint 2
+
+| Sprint | |
+|--------|----------------|
+| Início | 21/11/2025 |
+| Término | 04/12/2025 |
+| Duração | 12 dias |
