+++ interpretador/parser/parser.y
@@ -1,0 +1,32 @@
+%{
+#include <stdio.h>
+#include <stdlib.h>
+
+int yylex(void);
+void yyerror(const char *s);
+%}
+
+%token NUM PLUS MINUS TIMES DIVIDE LPAREN RPAREN
+
+
+%%
+
+expressao:
+    expressao PLUS expressao
+  | expressao MINUS expressao
+  | expressao TIMES expressao
+  | expressao DIVIDE expressao
+  | LPAREN expressao RPAREN
+  | NUM
+  ;
+
+%%
+
+void yyerror(const char *s) {
+    fprintf(stderr, "Erro sintático: %s\n", s);
+}
+
+int main(void) {
+    yyparse();
+    return 0;
+}
