--- lexer/lexer.l
@@ -1,17 +1,0 @@
-%{
-#include "parser.tab.h"
-%}
-
-%%
-
-[0-9]+          { yylval = atoi(yytext); return NUM; }
-"+"             { return PLUS; }
-"-"             { return MINUS; }
-"*"             { return TIMES; }
-"/"             { return DIVIDE; }
-"("             { return LPAREN; }
-")"             { return RPAREN; }
-[ \t\n]+        { /* Ignorar espaços em branco */ }
-.               { printf("Caractere inválido: %s\n", yytext); }
-
-%%
